Set MUNGE_ROOT for Travis CI build
The Travis CI build began failing a while back.  Commits that
had previously succeeded would now fail when re-tested.  This was
traced down to overly-permissive permissions now appearing on the
home directory (/home/travis/build/dun/munge).  This was presumably
caused by an update to the Travis CI image.

Sharness tests have their HOME set to the SHARNESS_TRASH_DIRECTORY.
By default, this is a subdirectory of the test directory (t).
When munged is started, it checks the security of its directories
including the parent directories on up to the root directory.
The now-permissive home directory causes these security checks to fail.

This commit fixes the Travis CI build by setting the new MUNGE_ROOT
variable to a subdirectory in "/tmp" (which has its sticky bit set).

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>