Reorder 0010-basic.t logfile check for replay
In t/0010-basic.t, a check is performed for a replayed credential.
This is immediately followed by a check for a replayed credential
log message.

On at least one occasion (on NetBSD 7.1.2, not reproducible), the
logfile check failed because the message had not yet appeared in the
logfile by the time the check was performed.  This happened despite the
logfile having its stream set to unbuffered (see log.c:log_open_file).

Check for a replayed credential log message after munged has stopped
to (hopefully) ensure all log messages have been written to file.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>